+++ frontend/src/components/ApplicationDetails.jsx
